requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
clerk-backend-api
python-dotenv==1.0.1
pydantic
ollama